Add concurrency
We run the individual search-replace operations asyncronously, but use a
buffered channel to ensure output remains in-order. Using a channel of
channels, we syncronously put a chan string placeholder in the channel
and send to it in a goroutine (async).

To ensure we don't get stuck in deadlock, we use a waitgroup to close
the channel (which effectively ends the master process) after all
channels have been processed.