Min replacement length is 2
The replacement value is not strictly necessary--we could support empty
replacements--but it adds some efficiency. The final step after
search-replace is fixing any serialized strings that were affected. If
we don't supply a replacement value, we have to look at every line. If
necessary, we may do that in the future, but for now the minimum will be
2.