Support multiple replacements
* Takes multiple sets of replacements
* Loops over replacements for each line

I was going to do all the string replacements and fix serialized strings
at the end, but the efficiency of the fixer function makes it moot
because we skip serialized strings that we didn't touch. Combined with
the slow performance of appending to a slice, it's actually faster like
this.