Fast conversion from []byte to string
fmt.Print needs a string, but now we're working with []byte. The normal,
memory safe way to do this is fairly slow and uses a lot of memory in
our use case because a new string needs to be allocated in memory.

Since we're not going to be operating on this given string anymore,
we just need to quickly write the contents to stdout. It doesn't matter
if the resulting string is immutable or not.